Fix a bug where identical transactions on the same day would result not result in additional transactions.
The `import_id` now takes occurences into account to prevent this issue by counting the number of times the row hash has been seen and including this occurence count in computing the actual hash.
This works similar to the rerun index which results in new hashes in order to support intentional re-importing.