Refactor the entire project to use classes.
This is a step towards making some upcoming changes easier. All import resposibilities are now handled in the Importer class. The Importer class delegates row transformations to the DataTransformer.
DataTransformer maintains the YNAB category IDs and produces clean rows for YNAB. DataTransformer caches the category IDs to be used in row transformations and manages import IDs of rows.
The RowInterpreter is now much easier to understand - this does the actual transformations of data fields from GoodBudget to YNAB.

In addition:
- Adds a new folder structure of a bin folder and a lib folder.
- Uses the new ES6 style functions.
- Makes heavy use of await for easy async.